requirements.txt
httpx>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
streamlit>=1.36.0
pandas>=2.2.0
plotly>=5.22.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
